implement (broken) neighbor algorithm on cpu
the problem is that I initially assumed that it would be ok for there to be hash collisions; as that would just result in traversing a superset of the intended neighbors. However, it is NOT ok because, if two cells hash to the same value, that means that any particles in those cells will be visited TWICE; once per each cell. Ultimately this leads to extra forces being accumulated. Instead, I think I need to give up and use a fixed-size spatial domain so that I can have a perfect cell position hash.