add support for uniform blocks in GLSL unused
I started working on this to add support for uniform buffers, but having done some limited GL profiling, I think this isn't important yet. The profiling, assuming it's accurate at all (with all the gl.finish calls) confirms my suspicions that sorting is the most expensive part of the update. The sort process makes lots of uniform changes in a tight loop, but it's already updating only the necessary ones.