hack to fix particle-particle collisions on cpu
before this change, adding the particle-particle collision impulse causes a distinct leftward drift of the particles. This is not present on GPU—the reason is that the GPU velocity array is (necessarily) double-buffered, while on the CPU the updateVelocity loop makes coherent updates to the velocity array, so we're reading our writes to velocity and getting weird behavior. This is a quick hack that just uses the velocity guess (which is not modified in this loop), and seems to work well enough. The right solution is double buffering velocity.