fix gpu sort for the billionth time
gpu mode had significantly more instability than cpu. because I saw small grid-like areas with artifacts in the simulation, I guessed that the neighbor iteration could be a problem. I started by replacing the foreachNeighbor macro with a simple loop over all particles, which resolved the artifacts/instability.This confirmed that the some part of the neighbors system was to blame. Then, I updated the CPU neighbors data structure to be exactly compatible with the GPU version. I used the CPU implementation to build the neighbors data structure, then copied it over to the GPU. This again resolved the problem, which confirmed that the neighbor traversal was working, but the data structure was not being built correctly. Finally, I used the GPU implementation to build the table and copied it to the CPU to inspect it. I noticed that some keys in the keyParticle array were not sorted, so I suspected a sorting bug. On a whim, I tested this change and found that it resolved the issue.