unify state, make a scene
this introduces the idea of scenes, which is a higher-level representation of an initial configuration of particles (positions, masses, phases, etc.). basically, it's backwards to set the number of particles ahead of time and then shoehorn the initial configuration into that, but that's basically what we have to do without the Scene concept, which can accumulate particles in a variably-sized array. now, I can just create particles as I please, and then convert those into an initial state. this will become particularly useful when I implement walls